feat(wallet): impement grpc handlers

+++ wallet/internal/adapters/db/sql/migrations/000005_currency.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE currency
 (
-  id   SMALLSERIAL PRIMARY KEY NOT NULL,
+  id   BIGSERIAL PRIMARY KEY NOT NULL,
   name VARCHAR(3)              NOT NULL -- USD, EGP, EUR, GBP, RUB
 );
 
@@ -12,7 +12,7 @@ VALUES ('USD'),
        ('RUB');
 
 ALTER TABLE accounts
-  ADD COLUMN currency_id SMALLINT NOT NULL default 1;
+  ADD COLUMN currency_id BIGINT NOT NULL default 1;
 
 ALTER TABLE accounts
   ADD FOREIGN KEY (currency_id) REFERENCES currency (id);
